requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
backcall==0.2.0
bleach==6.2.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
Django==5.2.7
django-user-agents==0.4.0
executing==2.2.0
fastjsonschema==2.21.1
gunicorn==23.0.0
idna==3.10
jedi==0.19.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.1
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pdfrw==0.4
pickleshare==0.7.5
platformdirs==4.3.6
prompt_toolkit==3.0.50
pure_eval==0.2.3
Pygments==2.19.1
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pywin32==308; sys_platform == 'win32'
pyzmq==26.2.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
urllib3==2.3.0
user-agents==2.2.0
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.8.2
yarg==0.1.9
cryptography
Faker==30.8.2
reportlab==4.2.5
